Use flow for button entries
