+++ azure-communication-ui/calling/src/main/java/com/azure/android/communication/ui/calling/models/CallCompositeExtensions.kt
@@ -3,6 +3,7 @@
 /* <CUSTOM_CALL_HEADER> */
 package com.azure.android.communication.ui.calling.models
 
+import android.content.Context
 import com.azure.android.communication.common.CommunicationIdentifier
 import com.azure.android.communication.ui.calling.presentation.manager.CallScreenInfoHeader
 
@@ -18,3 +19,17 @@ internal fun CallCompositeCallScreenHeaderOptions.setManager(manager: CallScreen
     this.callScreenInfoHeader = manager
 }
 /* </CUSTOM_CALL_HEADER> */
+
+internal fun createButtonClickEvent(
+    context: Context,
+    buttonOptions: CallCompositeButtonOptions,
+): CallCompositeButtonClickEvent {
+    return CallCompositeButtonClickEvent(context, buttonOptions)
+}
+
+internal fun createCustomButtonClickEvent(
+    context: Context,
+    buttonOptions: CallCompositeCustomButtonOptions,
+): CallCompositeCustomButtonClickEvent {
+    return CallCompositeCustomButtonClickEvent(context, buttonOptions)
+}
