+++ azure-communication-ui/calling/src/main/java/com/azure/android/communication/ui/calling/presentation/fragment/calling/CallingViewModel.kt
@@ -189,6 +189,7 @@ internal class CallingViewModel(
         captionsLanguageSelectionListViewModel.init(state.captionsState)
         captionsLayoutViewModel.init(state.captionsState, state.visibilityState)
 
+        moreCallOptionsListViewModel.init(state.visibilityState)
         super.init(coroutineScope)
     }
